requirements.txt
awscli==1.20.40
azure-ai-textanalytics==5.2.0b1
azure-common==1.1.27
azure-core==1.18.0
boto3==1.18.37
botocore==1.21.40
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.5
colorama==0.4.3
docutils==0.15.2
google-api-core==2.0.1
google-auth==2.1.0
google-cloud-language==2.2.2
googleapis-common-protos==1.53.0
grpcio==1.40.0
ibm-cloud-sdk-core==3.11.3
ibm-watson==5.2.0
idna==3.2
isodate==0.6.0
jmespath==0.10.0
msrest==0.6.21
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.3
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
websocket-client==0.48.0
